requirements.txt
asgiref==3.5.0
astroid==2.9.3
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.2
django-cors-headers==3.11.0
django-extensions==3.1.5
django-filter==21.1
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
filelock==3.4.2
gunicorn==20.1.0
identify==2.4.9
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.0
pre-commit==2.17.0
PyJWT==2.3.0
pylint==2.12.2
pylint-django==2.5.0
pylint-plugin-utils==0.7
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing-extensions==4.1.0
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
wrapt==1.13.3
